+++ src/LogIn (1).html
@@ -1,0 +1,58 @@
+<!DOCTYPE html>
+<html lang="en" dir="ltr">
+  <head>
+    <meta charset="utf-8">
+    <title>Contact</title> 
+    
+    
+    <link rel="stylesheet" href="css/landing.css">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+
+    
+    </head>
+  <body>
+    <div class="logo">
+      <img src="img/Contact.png" width=150 alt="contact logo">
+    </div>
+     
+    <div class="wrapper">
+      <div class="title-text">
+        <div class="title login">
+         
+            Login Form</div>
+            <div class="title signup">
+            Signup Form
+          </div>
+        </div>
+    <div class="hand">
+        <img src="img/Mask-Group-1.png" width="120">
+    </div>
+    <div class="form-container">
+       
+    </div>
+        <div class="form-inner">
+            <form action="#" class="login">
+                <div class="field">
+                    <input type="text" placeholder="Username" id="logUsername" required>
+                </div>
+        <div class="field">
+            <input type="password" placeholder="Password" id="logPass" required>
+        </div>
+            <div class="pass-link">
+
+                <div class="field btn">
+                    <div class="btn-layer">
+                        
+                    </div>
+                    <input type="submit" value="Login" id="logbutton" >
+                </div>
+                    <div class="signup-link">
+                      <div id="error"></div>
+Not a member? <a href="Register.html">Signup now</a>
+                </div>
+                </form>
+
+                </div>
+                <script type="text/javascript" src="js/login.js"></script>
+  </body>
+</html>
