+++ src/api/signup.php
