+++ src/js/register.js
